refactor: LatestArticles section
